Configured the AWS S3 image upload services

+++ src/main/java/com/Anshul/FoodDeliveryApp/service/FoodServiceImpl.java
@@ -6,13 +6,23 @@ import com.Anshul.FoodDeliveryApp.io.FoodResponse;
 import com.Anshul.FoodDeliveryApp.repository.FoodRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
+import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
+import software.amazon.awssdk.services.s3.model.DeleteObjectResponse;
+import software.amazon.awssdk.services.s3.model.PutObjectRequest;
+import software.amazon.awssdk.services.s3.model.PutObjectResponse;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
+
 
 @Service
 public class FoodServiceImpl implements FoodService {
@@ -22,9 +32,39 @@ public class FoodServiceImpl implements FoodService {
 	@Autowired
 	private FoodRepository foodRepo;
 
+	@Autowired
+	private S3Client s3Client;
+
+	@Value( "${aws.s3.bucketname}")
+	private String bucketName;
+
 	@Override
 	public String uploadFile(MultipartFile file) {
-		return "url";
+
+		// get the file extension
+		String filenameExtension = file.getOriginalFilename().substring(file.getOriginalFilename().lastIndexOf(".") + 1);
+
+		// generate unique
+		String key = UUID.randomUUID().toString()+"."+filenameExtension;
+
+		try{
+			PutObjectRequest putObjectRequest = PutObjectRequest.builder()
+					.bucket(bucketName)
+					.key(key)
+					.acl("public-read")
+					.contentType(file.getContentType())
+					.build();
+
+			PutObjectResponse response = s3Client.putObject(putObjectRequest, RequestBody.fromBytes(file.getBytes()));
+
+			if(response.sdkHttpResponse().isSuccessful()){
+				return "https://"+bucketName+".s3.amazonaws.com/"+key;
+			}else{
+				throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "File upload failed");
+			}
+		}catch (IOException e) {
+			throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Error while uploading file to S3 bucket");
+		}
 	}
 
 	// add food
@@ -66,7 +106,13 @@ public class FoodServiceImpl implements FoodService {
 
 	@Override
 	public Boolean deleteFile(String filename) {
-		return null;
+		DeleteObjectRequest deleteObjectRequest = DeleteObjectRequest.builder()
+				.bucket(bucketName)
+				.key(filename)
+				.build();
+
+		s3Client.deleteObject(deleteObjectRequest);
+		return true;
 	}
 
 	@Override
@@ -78,12 +124,17 @@ public class FoodServiceImpl implements FoodService {
 		String imageurl = response.getImageUrl();
 
 		// get the filename from the image url
+		String filename = imageurl.substring(imageurl.lastIndexOf("/") + 1);
 
 		// delete the image from aws
+		Boolean isDeleted = deleteFile(filename);
 
-		// if deleted successfully, then only,
-
-		foodRepo.deleteById(response.getId());
+		// if deleted successfully, then only delete the food from the database
+		if(
+				isDeleted
+		){
+			foodRepo.deleteById(response.getId());
+		}
 	}
 
 	// Convert FoodRequest to FoodEntity
